Refactor SidebarOpener
