+++ src/pages/Admin/Sidebar/SidebarOpener/index.ts
@@ -1,0 +1,1 @@
+export * from "./SidebarOpener";
